Make editor group navigation customizable
