+++ packages/plugin-ext-vscode/src/browser/plugin-vscode-commands-contribution.spec.ts
@@ -14,20 +14,37 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
 // *****************************************************************************
 
+import { enableJSDOM } from '@theia/core/lib/browser/test/jsdom';
+
+const disableJSDOM = enableJSDOM();
+// xterm.js is loaded transitively by the command contribution and probes canvas
+// support at module-load time. These tests do not render a terminal.
+const canvasProto = (globalThis as { HTMLCanvasElement?: { prototype: { getContext?: unknown } } }).HTMLCanvasElement?.prototype;
+if (canvasProto) {
+    canvasProto.getContext = () => undefined;
+}
+
+import { FrontendApplicationConfigProvider } from '@theia/core/lib/browser/frontend-application-config-provider';
+FrontendApplicationConfigProvider.set({});
+
 import { expect } from 'chai';
-import { findClosestEditorGroup, EditorGroupNavigationRect } from './editor-group-navigation';
+import { EditorGroupNavigationRect, PluginVscodeCommandsContribution } from './plugin-vscode-commands-contribution';
+
+after(() => disableJSDOM());
 
 function rect(left: number, top: number, width: number, height: number): EditorGroupNavigationRect {
     return { left, top, right: left + width, bottom: top + height, width, height };
 }
 
-describe('EditorGroupNavigation', () => {
+describe('PluginVscodeCommandsContribution', () => {
+    const contribution = new PluginVscodeCommandsContribution();
+
     describe('findClosestEditorGroup', () => {
         it('prefers the orthogonally-overlapping group when navigating up out of a split column', () => {
             const left = rect(0, 0, 50, 100);
             const upperRight = rect(50, 0, 50, 50);
             const lowerRight = rect(50, 50, 50, 50);
-            expect(findClosestEditorGroup(lowerRight, [left, upperRight], 'up')).to.equal(1);
+            expect(contribution['findClosestEditorGroup'](lowerRight, [left, upperRight], 'up')).to.equal(1);
         });
 
         it('navigates between columns in a 2x2 grid without skipping', () => {
@@ -36,16 +53,16 @@ describe('EditorGroupNavigation', () => {
             const bottomLeft = rect(0, 50, 50, 50);
             const bottomRight = rect(50, 50, 50, 50);
             const candidates = [topRight, bottomLeft, bottomRight];
-            expect(findClosestEditorGroup(topLeft, candidates, 'right')).to.equal(0);
-            expect(findClosestEditorGroup(topLeft, candidates, 'down')).to.equal(1);
+            expect(contribution['findClosestEditorGroup'](topLeft, candidates, 'right')).to.equal(0);
+            expect(contribution['findClosestEditorGroup'](topLeft, candidates, 'down')).to.equal(1);
         });
 
         it('returns -1 when no candidate lies in the requested direction', () => {
             const left = rect(0, 0, 50, 100);
             const right = rect(50, 0, 50, 100);
-            expect(findClosestEditorGroup(left, [right], 'left')).to.equal(-1);
-            expect(findClosestEditorGroup(left, [right], 'up')).to.equal(-1);
-            expect(findClosestEditorGroup(left, [], 'right')).to.equal(-1);
+            expect(contribution['findClosestEditorGroup'](left, [right], 'left')).to.equal(-1);
+            expect(contribution['findClosestEditorGroup'](left, [right], 'up')).to.equal(-1);
+            expect(contribution['findClosestEditorGroup'](left, [], 'right')).to.equal(-1);
         });
     });
 });
